refactor: 릴리즈 노트 단건 조회 페이지 분리
